Heroku Nueva Versión

See merge request invalidexception1/sixtysecondsserver!8

+++ app/Http/Controllers/PostController.php
@@ -13,8 +13,11 @@ use App\Services\FileSystem;
 use Response;
 use DB;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Storage;
+
 class PostController extends Controller
 {
+  
     protected CommentChainService $commentChainService;
     protected FileSystem $fileSystem;
     public function __construct(CommentChainService $_commentChainService, FileSystem $_fileSystem)
@@ -55,44 +58,45 @@ class PostController extends Controller
      */
     public function store(StorePostRequest $request)
     {
-        Log::channel('stderr')->info("Inserción de datos");
-        Log::channel('stderr')->info($request);
-       
+        //Log::channel('stderr')->info("Inserción de datos");
+        //Log::channel('stderr')->info($request);
+
         $response = [];
         try{
             DB::beginTransaction();
             $post = new Post();
             //$post->idPost = $request->id;
-            $post->idUsuario = $request->userId;
+            $post->idUsuario = $request->user()->idUsuario;
             $visibilidad = Visibilidad::where('Visibilidad', $request->visibility )->first();
-            Log::channel('stderr')->info('Visibilidad'.$visibilidad);
+
             if(!isset($visibilidad) || empty($visibilidad))
             {
                 $visibilidad = Visibilidad::where('Visibilidad','Public')->first();
             }
+       
             $audio = $this->fileSystem->saveFile($request->file('file'),$visibilidad->idVisibilidad);
             $post->idAudio = $audio->idAudio;
             $post->Titulo = $request->title;
-            
+
             $post->idVisibilidad = $visibilidad->idVisibilidad;
             $post->FechaCreacion = $request->creationDate;
             $post->Likes = 0;
             $post->Dislikes = 0;
-            
+
             if($post->save())
             {
                 if(isset($request->category) && !empty($request->category))
                 {
                     $tematicas = explode(',',$request->category);
-                    Log::channel('stderr')->info('Guardado Correcto');
+                    //Log::channel('stderr')->info('Guardado Correcto');
                     foreach($tematicas as $tematica)
                     {
                         $tematica = Tematica::where('Nombre', $tematica)->first();
-                    
+
                         if(isset($tematica))
                             $post->Tematica()->attach($tematica->idTematica);
                         //$user->roles()->attach($roleId, ['expires' => $expires]);
-                    } 
+                    }
                 }
                 $response = [ 'id' => $post->idPost];
 
@@ -101,18 +105,18 @@ class PostController extends Controller
             {
                 throw new Exception ('Error al almacenar el post');
             }
-            
-            
+
+
             DB::commit();
        }
         catch(Exception $e){
             DB::rollback();
-            $e->item = $request;
+            
             return  Response::json($e, 500); // Status code here
 
         }
         return Response::json($response, 201); // Status code here
-       
+
     }
 
     /**
@@ -137,7 +141,7 @@ class PostController extends Controller
         return Response::json($response, $returnCode); // Status code here
 
     }
- 
+
 
 
     /**
@@ -151,45 +155,53 @@ class PostController extends Controller
     {
         $response = [];
         try{
-            
+
             DB::beginTransaction();
-            $post = Post::where('idPost',$request->id)->first();
+   
+            $post = Post::where('idPost',$request->post_id)->first();
             //$post->idPost = $request->id;
-           
-            $visibilidad = Visibilidad::where('Visibilidad', $request->visibility )->first();
-            Log::channel('stderr')->info('Visibilidad'.$visibilidad);
+
+                if($post->Usuario->idUsuario != $request->user()->idUsuario)
+                {
+                    DB::rollback();
+                    $response[] = ['message' => 'Error: No estas autorizado para editar este post.'];
+                    return  Response::json($response, 401); // Status code here
+                }
+            if($request->visibility)
+                $visibilidad = Visibilidad::where('Visibilidad', $request->visibility )->first();
+
             if(!isset($visibilidad) || empty($visibilidad))
             {
-                $visibilidad = $post->Visibilidad;
+                $visibilidad = $post->Visibilidad ;
             }
-            // Se podrá cambiar el audio? 
+            // Se podrá cambiar el audio?
             //$audio = $this->fileSystem->saveFile($request->file('file'),$visibilidad->idVisibilidad);
             //$post->idAudio = $audio->idAudio;
             if(isset($request->title))
                 $post->Titulo = $request->title;
-            
+
             $post->idVisibilidad = $visibilidad->idVisibilidad;
             //$post->FechaCreacion = $request->creationDate;
             //$post->Likes = 0;
             //$post->Dislikes = 0;
-            
+
             if($post->save())
             {
                 if(isset($request->category))
                 {
                     $tematicasNewNombre = explode(',',$request->category);
-                    
+
                     $tematicasOld = [];
                     $tematicasNew =  Tematica::whereIn('Nombre',$tematicasNewNombre)->pluck('idTematica')->toArray();
                     $tematicasOld = $post->Tematica()->get()->pluck('idTematica')->toArray();
-                    
+
                     $tematicasAEliminar = array_diff($tematicasOld,$tematicasNew);
                     $tematicasAInsertar = array_diff($tematicasNew,$tematicasOld);
 
                     $post->Tematica()->detach($tematicasAEliminar);
                     $post->Tematica()->attach($tematicasAInsertar);
-         
-         
+
+
                 }
                 $response = [ 'id' => $post->idPost];
 
@@ -198,13 +210,13 @@ class PostController extends Controller
             {
                 throw new Exception ('Error al almacenar el post');
             }
-            
-            
+
+
             DB::commit();
        }
         catch(Exception $e){
             DB::rollback();
-            $e->item = $request;
+            
             return  Response::json($e, 500); // Status code here
         }
         return Response::json($response, 200); // Status code here
@@ -216,34 +228,44 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\Response
      */
-    public function destroy($idPost)
+    public function destroy(Request $request ,$idPost)
     {
         $response = [];
         try{
-            
+
             DB::beginTransaction();
             $post = Post::where('idPost',$idPost)->first();
-            if($post->exists())
+                             
+            if($post != null && $post->exists() )
             {
+                if($post->Usuario->idUsuario != $request->user()->idUsuario)
+                {
+                    DB::rollback();
+                    $response[] = ['message' => 'Error: No estas autorizado para eliminar este post.'];
+                    return  Response::json($response, 401); // Status code here
+                }
                 if($post->Tematica()->exists())
                     $post->Tematica()->detach($post->Tematica()->get()->pluck('idTematica')->toArray());
-               
+
                 if( $post->Comentarios != null)
                 {
-                    
+
                     $comentarios = $post->Comentarios()
                                         ->whereNotNull('idComentarioPadre')
                                         ->orderBy('idComentarioPadre', 'desc')
                                         ->get();
                                  // ->get();
-                    Log::channel('stderr')->info('Comentarios'.json_encode($comentarios));
+                    //Log::channel('stderr')->info('Comentarios'.json_encode($comentarios));
                     foreach($comentarios as $comentario)
                     {
                         $comentario->delete();
                     }
                     $post->Comentarios()->delete();
                 }
+              
                 $post->delete();
+                if($post->Audio()->exists())
+                    $post->Audio->delete();
                 $response[] = ['message' => 'Post eliminado correctamente'];
             }
             else{
@@ -255,18 +277,107 @@ class PostController extends Controller
         }
         catch(Exception $e){
             DB::rollback();
-            $e->item = $request;
             return  Response::json($e, 500); // Status code here
         }
         return Response::json($response, 200); // Status code here
     }
     public function addComment(Request $request)
     {
+        $response = [];
+        try{
+            
+            DB::beginTransaction();
+            $comentario = new Comentario();
+            $comentario->idUsuario = $request->user()->idUsuario;
+            $comentario->FechaCreacion = $request->creationDate;
+            $comentario->idPost = $request->postID;
+            $visibilidad = Visibilidad::where('Visibilidad', $request->visibility )->first();
+                    
+            if(!isset($visibilidad) || empty($visibilidad))
+            {
+                $visibilidad = $post->Visibilidad;
+            }
 
+            $audio = $this->fileSystem->saveFile($request->file('file'),$visibilidad->idVisibilidad);
+            $comentario->idAudio = $audio->idAudio;
+            DB::commit();
+            $comentario->Likes = 0;
+            $comentario->Dislikes = 0;
+            $comentario->idComentarioPadre = $request->parentId;
+
+            if($comentario->save())
+            {
+                
+                $response = [ 'id' => $comentario->idComentario];
+
+            }
+            else
+            {
+                throw new Exception ('Error al almacenar el post');
+            }
+            DB::commit();
+            
+        }
+        catch(Exception $e){
+            DB::rollback();
+            
+            return  Response::json($e, 500); // Status code here
+        }
+        return Response::json($response, 200); // Status code here
+       
     }
-    public function destroyComment(Request $request)
+    public function destroyComment(Request $request,$idComment)
     {
+         $response = [];
+        try{
+            
+            DB::beginTransaction();
+            $comentario = Comentario::where('idcomentario',$idComment)->first();
+            
+            if($comentario != null && $comentario->exists())
+            {
+                
+                 if($comentario->Usuario->idUsuario != $request->user()->idUsuario)
+                {
+                    DB::rollback();
+                    $response[] = ['message' => 'Error: No estas autorizado para eliminar este comentario.'];
+                    return  Response::json($response, 401); // Status code here
+                }
+                 
+                if( $comentario->Comentarios != null)
+                {
+                    this->deleteCommentChildren($comentario->Comentarios);
+                }
+                if($post->Audio()->exists())
+                    $post->Audio->delete();
+                $comentario->delete();
+                $response[] = ['message' => 'Post eliminado correctamente'];
+            }
+            else{
+                $response[] = ['message' => 'El post a eliminar no existe'];
 
-    }
+            }
+            DB::commit();
+
+        }
+        catch(Exception $e){
+            DB::rollback();
+            
+            return  Response::json($e, 500); // Status code here
+        }
+        return Response::json($response, 200); // Status code here
     
+    }
+    protected function deleteCommentChildren($comentarios)
+    {
+        foreach($comentarios as $comentario)
+        {
+            if($comentario->Comentarios != null)
+            {
+                this->deleteCommentChildren($comentario->Comentarios());
+            }
+            $comentario->delete();
+        }
+    }
+
 }
